% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianWinRatio-package.R
\docType{package}
\name{BayesianWinRatio-package}
\alias{BayesianWinRatio}
\alias{BayesianWinRatio-package}
\title{BayesianWinRatio: Performing Bayesian Futility Monitoring Using the Win Ratio Approach for Multiple Time-to-event Endpoints}
\description{
This package performs Bayesian futility monitoring using the win ratio approach based on the manuscript "The Win Ratio Approach in Bayesian Monitoring for Two-Arm Phase II Clinical Trial Designs with Multiple Time-to-Event Endpoints".
}
\author{
\strong{Maintainer}: Xinran Huang \email{huangxr.2019@gmail.com} (\href{https://orcid.org/0000-0002-8126-9735}{ORCID})

}
\keyword{internal}
