% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compare}
\alias{compare}
\title{compare}
\usage{
compare(
  M_iter,
  n_current_ctrl,
  n_current_trt,
  postData,
  time_trt,
  time_ctrl,
  surv_trt,
  surv_ctrl
)
}
\arguments{
\item{M_iter}{Number of iterations after burning and thinning.}

\item{n_current_ctrl}{Sample size at a interim analysis point for the control arm.}

\item{n_current_trt}{Sample size at a interim analysis point for the treatment arm.}

\item{postData}{A cube with n rows, two columns and N_iter slices. Slice v, v = 1, ..., N_iter, contains n pairs of event times sampled from bivariate lognormal distribution with mean MU and variance Sigma, where
Mu and Sigma are generated from the posterior normal and inverse-wishart distribution, respectively, from the v-th iteration of the gibbs sampler.}

\item{time_trt}{A field containing M_iter elements, each of which is a vector of length n_current_trt. These vectors store the censoring time generated from the estimated censoring distribution from the treatment arm.}

\item{time_ctrl}{A field containing M_iter elements, each of which is a vector of length n_current_ctrl. These vectors store the censoring time generated from the estimated censoring distribution from the control arm.}

\item{surv_trt}{A field containing M_iter elements, each of which is a vector of length n_current_trt. These vectors store the survival probabilities generated from the estimated censoring distribution from the treatment arm.}

\item{surv_ctrl}{A field containing M_iter elements, each of which is a vector of length n_current_ctrl. These vectors store the survival probabilities generated from the estimated censoring distribution from the control arm.}
}
\value{
A list of WR. WR is a matrix with one row and M_iter columns. Each row contains the estimated win ratio for each iteration.
}
\description{
This is an interim function used within the winratio function that calculates the IPCW-adjusted win ratio following Equations (1) - (3).
}
