% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation_final.R
\name{data.simulation}
\alias{data.simulation}
\title{data.simulation}
\usage{
data.simulation(
  N.sim,
  N.max,
  mu.trt,
  Sigma.trt,
  mu.ctrl,
  Sigma.ctrl,
  cens_upper
)
}
\arguments{
\item{N.sim}{Number of simulations.}

\item{N.max}{Maximum number of patients to enroll.}

\item{mu.trt}{A vector containing the mean time to each event (logarithm) for the treatment arm.}

\item{Sigma.trt}{A vector representing the variance-covariance matrix of the time to each event (logarithm) for the treatment arm.}

\item{mu.ctrl}{A vector containing the mean time to each event (logarithm) for the control arm.}

\item{Sigma.ctrl}{A vector representing the variance-covariance matrix of the time to each event (logarithm) for the control arm.}

\item{cens_upper}{Upper limit for the censoring time, assuming that the censoring time is generated from Uniform(0, cens_upper).}
}
\value{
An array with N.max rows, 7 columns and N.sim slices. The columns are:\tabular{ll}{
\code{recurrence_t} \tab Time to recurrence.  \cr
\tab \cr
\code{death_t} \tab Time to death. \cr
\tab \cr
\code{censor_t} \tab Time to censor. \cr
\tab \cr
\code{group} \tab Treatment assignment.  \cr
\tab \cr
\code{id} \tab Patient id, numeric starting from 1. \cr
\tab \cr
\code{delta1} \tab Placeholder for the censoring indicator for recurrence, leave blank. \cr
\tab \cr
\code{delta2} \tab Placeholder for the censoring indicator for death, leave blank. \cr
}
}
\description{
This function simulates data.
}
\examples{
\dontrun{
data.simulation(N.sim = 1000, N.max = 20,
               mu.trt = c(0.2,0.3), Sigma.trt = matrix(c(1,0.5,0.5,1), nrow = 2, byrow = T),
               mu.ctrl = c(0.2,0.3),Sigma.ctrl = matrix(c(1,0.5,0.5,1), nrow = 2, byrow = T),
               cens_upper = 5)
}
}
