% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_ratio_final.R
\name{est_cens_time}
\alias{est_cens_time}
\title{est_cens_time}
\usage{
est_cens_time(fit, n)
}
\arguments{
\item{fit}{Estimated censoring distribution.}

\item{n}{A numeric value indicating number of censoring time to generate. Either n.current.trt or n.current.ctrl.}
}
\value{
Censoring time for each subject.
}
\description{
This function is an intermediate function used inside the winratio function to generate censoring time from estimated censoring distribution.
}
\examples{
\dontrun{est_cens_time(trt_cens.est,n.current.trt)}
}
