% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_ratio_final.R
\name{make_pos_def}
\alias{make_pos_def}
\title{make_pos_def}
\usage{
make_pos_def(mat, eps = 1e-06)
}
\arguments{
\item{mat}{Sample covariance matrix}

\item{eps}{a small value to ensure the matrix is positive-definite}
}
\value{
A positive definite matrix.
}
\description{
This function is an intermediate function used inside the winratio function to ensure the sample covariance matrix is positive definite
}
\examples{
\dontrun{make_pos_def(temp)}  
}
