% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test}
\alias{test}
\title{Test data}
\format{
\subsection{\code{test}}{

A dataframe of 100 rows and 5 columns:
\describe{
\item{id}{Patient ID}
\item{arm}{A numeric value representing treatment arms. 1 = control, 2 = treatment}
\item{recurrence}{Time to recurrence: NA if not observed}
\item{death}{Time to death: NA if not observed}
\item{censor}{Time to last follow-up}
}
}
}
\usage{
test
}
\description{
A test data used for demo.
}
\keyword{datasets}
