% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_theta}
\alias{update_theta}
\title{update_theta}
\usage{
update_theta(N_iter, dd, n, m0, L0, S0, v0, time_max, init_Sigma)
}
\arguments{
\item{N_iter}{Number of iterations for gibbs sampler.}

\item{dd}{The data at a interim analysis point for either treatment or control arm. Either use currentData.trt or currentData.ctrl.}

\item{n}{Sample size at a interim analysis point for either treatment or control arm. Either use n.current.trt or n.current.ctrl.}

\item{m0}{Prior mean for mu.trt/mu.ctrl.}

\item{L0}{Prior covariance for mu.trt/mu.ctrl.}

\item{S0}{Prior scale matrix for Sigma.trt/Sigma.ctrl.}

\item{v0}{Prior degrees of freedom for Sigma.trt/Sigma.ctrl.}

\item{time_max}{The upper limit for the recurrence and death time sampled from truncated normal. This will set the upper limit to to time_max rather than Inf.}

\item{init_Sigma}{An initial Sigma covariance matrix.}
}
\value{
A list with the following components:\tabular{ll}{
\code{MU} \tab A matrix with N_iter rows and two columns. Row v, v = 1, ..., N_iter, contains posterior mean vector generated from the v-th iteration of the gibbs sampler.  \cr
\tab \cr
\code{post_data} \tab A cube with n rows, two columns and N_iter slices. Slice v, v = 1, ..., N_iter, contains n pairs of event times sampled from bivariate lognormal distribution with mean MU and variance Sigma, where
Mu and Sigma are generated from the posterior normal and inverse-wishart distribution, respectively, from the v-th iteration of the gibbs sampler.  \cr
\tab \cr
\code{Sigma} \tab A cube with two rows, two columns and N_iter slices. Slice v, v = 1, ..., N_iter, contains posterior covariance matrix generated from the v-th iteration of the gibbs sampler.   \cr
}
}
\description{
This is an interim function used within the winratio function that performs Gibbs sampler and returns posterior mean, covariance matrix, and event times for the proposed design, following Supplementary file S1.
}
