% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_theta_univariate}
\alias{update_theta_univariate}
\title{update_theta_univariate}
\usage{
update_theta_univariate(N_iter, dd, n, L0, m0, v0, S0, time_max, init_Sigma)
}
\arguments{
\item{N_iter}{Number of iterations for gibbs sampler.}

\item{dd}{The data at a interim analysis point for either treatment or control arm. Either use currentData.trt or currentData.ctrl.}

\item{n}{Sample size at a interim analysis point for either treatment or control arm. Either use n.current.trt or n.current.ctrl.}

\item{L0}{Prior variance for mu.trt/mu.ctrl.}

\item{m0}{Prior mean for mu.trt/mu.ctrl.}

\item{v0}{v0/2 is the prior shape for Sigma.trt/Sigma.ctrl.}

\item{S0}{v0*S0/2 is the prior scale for Sigma.trt/Sigma.ctrl.}

\item{time_max}{The upper limit for the recurrence and death time sampled from truncated normal. This will set the upper limit to to time_max rather than Inf.}

\item{init_Sigma}{A numeric value represents the time to event (logarithm) for the control arm.}
}
\value{
A list of theta. theta is A matrix with N_iter rows and two columns. Row v, v = 1, ..., N_iter, contains posterior mean vector generated from the v-th iteration of the gibbs sampler.
}
\description{
This function is an interim function used within the OCC.Table function that performs Gibbs sampler and returns posterior mean, variance, and event times (log) for the conventional designs, following Supplementary file S1.
}
