% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_ratio_final.R
\name{winratio}
\alias{winratio}
\title{winratio}
\usage{
winratio(
  currentdd,
  n_current,
  Time_current,
  Time_entry,
  m0,
  L0,
  v0,
  S0,
  time_max,
  eta,
  lambda,
  N.max,
  thin_MCMC,
  Niter
)
}
\arguments{
\item{currentdd}{The data at a interim analysis point.}

\item{n_current}{Sample size at a interim analysis point.}

\item{Time_current}{Time at a interim analysis point.}

\item{Time_entry}{Time at entry for each patient at a interim analysis point.}

\item{m0}{Prior mean for mu.}

\item{L0}{Prior covariance for mu.}

\item{v0}{Prior degrees of freedom for Sigma.}

\item{S0}{Prior scale matrix for Sigma.}

\item{time_max}{The upper limit for the recurrence and death time sampled from truncated normal. This will set the upper limit to to time_max rather than Inf.}

\item{eta}{A pre-specified lower bound of acceptable performance based on historical information.}

\item{lambda}{Cutoff parameter.}

\item{N.max}{Maximum number of patients to enroll.}

\item{thin_MCMC}{Thinning degree.}

\item{Niter}{Number of iterations for gibbs sampler.}
}
\value{
A list with the following components:\tabular{ll}{
\code{probs} \tab A numeric value, which is the posterior probability of \eqn{(\widehat{WR} > eta)}.  \cr
\tab \cr
\code{cutoff} \tab A numeric value. cutoff = (n_current/N.max)**lambda.  \cr
\tab \cr
\code{WR} \tab A numeric value. WR = the average posterior win ratio.  \cr
}
}
\description{
This function is an intermediate function used inside the OCC.Table function to obtain the posterior probability of (WR > eta).
}
\examples{
\dontrun{
winratio(currentdd = currentData,n_current = n.current, Time_current = Time.current,
Time_entry = Time.entry,m0 = c(0,0),L0 = diag(10^6, 2),v0 = 4,S0 = diag(10^(-6), 2),
time_max = 20,eta = 1,lambda = 0.25,N.max = 100, thin_MCMC = 5,
Niter = 100000)
} 
}
